CAR-20 правки после ревью
