+++ packages/client/src/pages/RatingPage/data.ts
@@ -1,25 +1,4 @@
-import type { TColumns, TDataSource } from './types'
-
-export const columns: TColumns = [
-  {
-    title: '',
-    dataIndex: 'num',
-    key: 'num',
-    width: 60,
-  },
-  {
-    title: 'Пользователь',
-    dataIndex: 'user',
-    ellipsis: true,
-    key: 'user',
-  },
-  {
-    title: 'Баллы',
-    dataIndex: 'scores',
-    key: 'scores',
-    ellipsis: true,
-  },
-]
+import type { TDataSource } from './types'
 
 export const dataSource: TDataSource = [
   {
