+++ packages/client/src/pages/RatingPage/index.tsx
@@ -1,7 +1,8 @@
 import { Card } from '@/shared/ui/Card'
 import { Table } from '@/shared/ui/Table'
 import { Content } from '@/shared/ui/Layout'
-import { columns, dataSource } from './data'
+import { dataSource } from './data'
+import { columns } from './utils/columns'
 
 const RatingPage = () => {
   return (
